Move `StashAccountAssetOps` to the common_strategies.rs
