+++ substrate/frame/derivatives/src/mock/mod.rs
@@ -68,8 +68,7 @@ use sp_runtime::{
 use xcm::prelude::*;
 use xcm_builder::{
 	unique_instances::{
-		ops::StashAccountAssetOps, ExtractAssetId, NonFungibleAsset, UniqueInstancesAdapter,
-		UniqueInstancesDepositAdapter,
+		ExtractAssetId, NonFungibleAsset, UniqueInstancesAdapter, UniqueInstancesDepositAdapter,
 	},
 	AllowUnpaidExecutionFrom, AsPrefixedGeneralIndex, FixedWeightBounds, MatchInClassInstances,
 	MatchedConvertedConcreteId, StartsWith,
